Use dense matrix factory everywhere in tests
